optimize code for posting zoom actions
