--- include/logger.h
@@ -1,12 +1,0 @@
-#pragma once
-
-void logger(
-	unsigned int,
-	std::tuple<std::string, std::string>
-);
-
-void logger(
-	unsigned int,
-	const std::string&,
-	const std::string&
-);
