+++ src/JS/JS.cpp
@@ -1,0 +1,14 @@
+#include "JS.h"
+
+std::string zoom_increment_js = {
+#include "zoom-increment.js"
+};
+
+std::string zoom_set_to_js = {
+#include "zoom-set-to.js"
+};
+
+std::string pass_key_bind_js = {
+#include "pass-key-bind.js"
+};
+
